Add directory client in components directory and move corresponding files
