+++ frontend/src/components/client/ClientRegisterForm.tsx
@@ -1,6 +1,6 @@
 import React, {useState} from "react"
-import {query} from "../libs/database";
-import Address from "../model/Address";
+import {query} from "../../libs/database";
+import Address from "../../model/Address";
 import Select from "react-select";
 import {observer} from "mobx-react";
 import {Result} from "surrealdb.js";
